Задание 5.12 "Модуляция"
