--- js/util.js
@@ -1,3 +1,0 @@
-// const VAR = 10;
-
-// export {VAR}
